Refactor #68 - move transaction package `core` to `domain.common`
