+++ app/src/main/java/com/nlab/reminder/domain/common/schedule/impl/DefaultModifyScheduleCompleteUseCase.kt
@@ -18,7 +18,7 @@ package com.nlab.reminder.domain.common.schedule.impl
 
 import com.nlab.reminder.core.kotlin.coroutine.Delay
 import com.nlab.reminder.core.kotlin.util.Result
-import com.nlab.reminder.core.util.transaction.TransactionIdGenerator
+import com.nlab.reminder.domain.common.util.transaction.TransactionIdGenerator
 import com.nlab.reminder.domain.common.schedule.*
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.firstOrNull
