+++ app/src/main/java/com/nlab/reminder/internal/common/di/ScheduleModule.kt
@@ -18,7 +18,7 @@ package com.nlab.reminder.internal.common.di
 
 import com.nlab.reminder.core.kotlin.coroutine.util.Delay
 import com.nlab.reminder.domain.common.util.link.LinkMetadataRepository
-import com.nlab.reminder.core.util.transaction.TransactionIdGenerator
+import com.nlab.reminder.domain.common.util.transaction.TransactionIdGenerator
 import com.nlab.reminder.domain.common.schedule.*
 import com.nlab.reminder.domain.common.schedule.impl.*
 import dagger.Module
